Make it work on local with docker compose

+++ sensor/src/main/java/org/tapsensors/sensor/SensorData.java
@@ -2,5 +2,5 @@ package org.tapsensors.sensor;
 
 import java.util.UUID;
 
-public record SensorData(UUID id, double temperature, double pressure) {
+public record SensorData(UUID id, int temperature, int pressure) {
 }
